refactor linkedin
