+++ packages/core/src/modules/linkedin.ts
@@ -1,236 +1,258 @@
-import dedent from 'dedent';
-import { sql, type Transaction } from 'kysely';
-import { match } from 'ts-pattern';
+import { sql, type Transaction, type Updateable } from 'kysely';
 import { z } from 'zod';
 
 import { type DB, db, point } from '@oyster/db';
-import { id } from '@oyster/utils';
+import { type Major } from '@oyster/types';
+import { id, run, splitArray } from '@oyster/utils';
 
-import { getChatCompletion } from '@/infrastructure/ai';
-import { job } from '@/infrastructure/bull';
 import { track } from '@/infrastructure/mixpanel';
-import { withCache } from '@/infrastructure/redis';
+import { cache, ONE_WEEK_IN_SECONDS, withCache } from '@/infrastructure/redis';
 import { runActor } from '@/modules/apify';
-import { LocationType } from '@/modules/employment/employment.types';
+import { saveSchoolIfNecessary } from '@/modules/education/use-cases/save-school-if-necessary';
+import {
+  EmploymentType,
+  LocationType,
+} from '@/modules/employment/employment.types';
 import { saveCompanyIfNecessary } from '@/modules/employment/use-cases/save-company-if-necessary';
 import {
   getAutocompletedCities,
   getCityDetails,
 } from '@/modules/location/location';
-import { STUDENT_PROFILE_URL } from '@/shared/env';
-import { ColorStackError } from '@/shared/errors';
 
-// Schemas
-
-const LinkedInDegreeType = z.enum([
-  'associate',
-  'bachelors',
-  'certificate',
-  'doctoral',
-  'masters',
-  'professional',
-]);
-
-const LinkedInEmploymentType = z.enum([
-  'apprenticeship',
-  'contract',
-  'freelance',
-  'full_time',
-  'internship',
-  'part_time',
-]);
-
-const LinkedInLocationType = z.enum(['hybrid', 'in_person', 'remote']);
-
-const LinkedInMajor = z.enum([
-  'artificial_intelligence',
-  'computer_science',
-  'data_science',
-  'electrical_or_computer_engineering',
-  'information_science',
-  'other',
-]);
-
-const LinkedInEducation = z.object({
-  degreeType: LinkedInDegreeType,
-  endDate: z.string().nullable(),
-  isSchoolOnLinkedIn: z.boolean(),
-  major: LinkedInMajor,
-  otherMajor: z.string().nullable(),
-  school: z.string(),
-  startDate: z.string().nullable(),
+const LinkedInDate = z.object({
+  month: z.string().nullish(),
+  text: z.string(),
+  year: z.number(),
 });
-
-type LinkedInEducation = z.infer<typeof LinkedInEducation>;
-
-const LinkedInExperience = z.object({
-  company: z.string(),
-  endDate: z.string().nullable(),
-  employmentType: LinkedInEmploymentType,
-  isCompanyOnLinkedIn: z.boolean(),
-  location: z.string().nullable(),
-  locationType: LinkedInLocationType.nullable(),
-  startDate: z.string(),
-  title: z.string(),
-});
-
-type LinkedInExperience = z.infer<typeof LinkedInExperience>;
 
 const LinkedInProfile = z.object({
-  educations: z.array(LinkedInEducation),
-  experiences: z.array(LinkedInExperience),
-  location: z.string().nullable(),
+  education: z.array(
+    z.object({
+      degree: z.string().nullish(),
+      endDate: LinkedInDate.nullish(),
+      fieldOfStudy: z.string().nullish(),
+      schoolName: z.string(),
+      schoolLinkedinUrl: z.string().url(),
+      startDate: LinkedInDate.nullish(),
+    })
+  ),
+  element: z.object({
+    headline: z.string().trim(),
+    location: z.object({
+      parsed: z.object({
+        text: z.string(),
+      }),
+    }),
+    photo: z.string().url().nullish(),
+  }),
+  experience: z.array(
+    z.object({
+      companyId: z.string().nullish(),
+      companyName: z.string(),
+      companyLinkedinUrl: z.string().url(),
+      description: z.string().nullish(),
+      employmentType: z
+        .enum([
+          'Apprenticeship',
+          'Contract',
+          'Freelance',
+          'Full-time',
+          'Internship',
+          'Part-time',
+          'Seasonal',
+          'Self-employed',
+          'Volunteer',
+        ])
+        .nullish(),
+      endDate: LinkedInDate.nullish(),
+      location: z.string().nullish(),
+      position: z.string(),
+      startDate: LinkedInDate.nullish(),
+      workplaceType: z.enum(['Hybrid', 'On-site', 'Remote']).nullish(),
+    })
+  ),
+  originalQuery: z.object({
+    query: z.string(),
+  }),
 });
 
 type LinkedInProfile = z.infer<typeof LinkedInProfile>;
+type LinkedInEducation = z.infer<typeof LinkedInProfile>['education'][number];
+type LinkedInExperience = z.infer<typeof LinkedInProfile>['experience'][number];
 
-const EducationChange = z.object({
-  data: LinkedInEducation.extend({ id: z.string().optional() }),
-  type: z.literal('education'),
-});
-
-const ExperienceChange = z.object({
-  data: LinkedInExperience.extend({ id: z.string().optional() }),
-  type: z.literal('experience'),
-});
-
-const LocationChange = z.object({
-  data: z.object({ location: z.string() }),
-  type: z.literal('location'),
-});
-
-const Change = z.discriminatedUnion('type', [
-  EducationChange,
-  ExperienceChange,
-  LocationChange,
-]);
-
-type Change = z.infer<typeof Change>;
-
-// Core
-
-/**
- * Syncs a member's LinkedIn profile history (work, education, etc.)
- * with their database record. This is a complex process which executes the
- * following steps:
- *
- * 1. Fetches the member's existing work/education history from the database.
- * 2. Fetches the member's LinkedIn profile data from Apify. This step also
- *    cleans the LinkedIn profile data to match our expected schema. This also
- *    removes any irrelevant data like high school education, service industry
- *    jobs, etc.
- * 3. Using AI, generates a list of differences between the member's LinkedIn
- *    profile data and their database record. This spits out a list of
- *    "changes" to either add or edit records in the database.
- * 4. Executes the changes to synchronize the database with the LinkedIn
- *    profile data.
- *
- * There are multiple potential failure points in this process, whether that
- * be with the LinkedIn profile scraping, the AI cleaning, the AI differential
- * generation or the database execution.
- */
-export async function syncLinkedInProfile(memberId: string): Promise<void> {
-  const [member, educations, workExperiences] = await db
+export async function syncLinkedInProfiles(memberIds?: string[]) {
+  const [members, educations, experiences] = await db
     .transaction()
     .execute(async (trx) => {
       return Promise.all([
-        getMemberForDifferential(trx, memberId),
-        getEducationHistoryForDifferential(trx, memberId),
-        getWorkHistoryForDifferential(trx, memberId),
+        getAllMembers(trx, memberIds),
+        getAllEducations(trx, memberIds),
+        getAllExperiences(trx, memberIds),
       ]);
     });
 
-  if (!member || !member.linkedInUrl) {
-    return;
-  }
+  const memberMap: Record<string, Member> = {};
+  const educationMap: Record<string, Education[]> = {};
+  const experienceMap: Record<string, Experience[]> = {};
 
-  // Scrapes the LinkedIn profile from the member's URL.
-  const linkedInProfile = await getLinkedInProfile(member.linkedInUrl);
+  members.forEach((member) => {
+    memberMap[member.linkedInUrl!] = member;
+  });
 
-  if (!linkedInProfile) {
-    console.log(`No LinkedIn profile found for ${member.linkedInUrl}.`);
-
-    return;
-  }
-
-  // Generates a list of changes to synchronize the database with the LinkedIn
-  // profile data.
-  const changes = await getProfileDifferential(
-    {
-      educations,
-      location: member.currentLocation,
-      workExperiences,
-    },
-    {
-      educations: linkedInProfile.educations,
-      location: linkedInProfile.location,
-      workExperiences: linkedInProfile.experiences,
+  educations.forEach((education) => {
+    if (educationMap[education.studentId]) {
+      educationMap[education.studentId].push(education);
+    } else {
+      educationMap[education.studentId] = [education];
     }
-  );
+  });
 
-  console.log(
-    `Found ${changes.length} changes for ${member.linkedInUrl}.`,
-    changes
-  );
+  experiences.forEach((experience) => {
+    if (experienceMap[experience.studentId]) {
+      experienceMap[experience.studentId].push(experience);
+    } else {
+      experienceMap[experience.studentId] = [experience];
+    }
+  });
 
-  if (!changes.length) {
-    return;
-  }
+  const batches = splitArray(members, 100);
 
-  // Executes the changes to synchronize the database with the LinkedIn.
-  await db.transaction().execute(async (trx) => {
-    const promises = changes.map((change) => {
-      return match(change)
-        .with({ type: 'education' }, async ({ data }) => {
-          return upsertEducationFromLinkedIn(trx, memberId, data);
-        })
-        .with({ type: 'experience' }, async ({ data }) => {
-          return upsertExperienceFromLinkedIn(trx, memberId, data);
-        })
-        .with({ type: 'location' }, async ({ data }) => {
-          return updateLocationFromLinkedIn(trx, memberId, data);
-        })
-        .exhaustive();
-    });
+  for (const batch of batches) {
+    const results: LinkedInProfile[] = [];
 
-    promises.push(
-      trx
-        .updateTable('students')
-        .set({ linkedinSyncedAt: new Date() })
-        .where('id', '=', memberId)
-        .execute()
+    // Need to filter through the batch and only get the ones that aren't
+    // already in the cache.
+    await Promise.all(
+      batch.map(async (member) => {
+        const profile = await cache.get<LinkedInProfile>(
+          `apify:harvestapi~linkedin-profile-scraper:${member.linkedInUrl}`
+        );
+
+        if (profile) {
+          results.push(profile);
+
+          const index = batch.findIndex(({ id }) => member.id === id);
+
+          if (index !== -1) {
+            batch.splice(index, 1);
+          }
+
+          return;
+        }
+      })
     );
 
-    await Promise.all(promises);
-  });
-
-  const educationChanges = changes.filter(({ type }) => type === 'education');
-  const experienceChanges = changes.filter(({ type }) => type === 'experience');
-  const locationChanges = changes.filter(({ type }) => type === 'location');
-
-  track({
-    event: 'LinkedIn Synced',
-    properties: {
-      '# of Education Changes': educationChanges.length,
-      '# of Work Experience Changes': experienceChanges.length,
-      'Location Changed': locationChanges.length > 0,
-    },
-    user: memberId,
-  });
-
-  if (member.slackId) {
-    job('notification.slack.send', {
-      channel: member.slackId,
-      message: `I synced your <${STUDENT_PROFILE_URL}/profile/work|work history>, <${STUDENT_PROFILE_URL}/profile/education|education history> and <${STUDENT_PROFILE_URL}/profile/general|current location> with your <${member.linkedInUrl}|LinkedIn>. Please make any changes if we got something wrong!`,
-      workspace: 'regular',
+    const newProfiles = await runActor({
+      actorId: 'harvestapi~linkedin-profile-scraper',
+      body: { urls: batch.map((member) => member.linkedInUrl!) },
+      schema: z.array(LinkedInProfile),
     });
+
+    results.concat(newProfiles);
+
+    await Promise.all(
+      results.map(async (profile) => {
+        await cache.set(
+          `apify:harvestapi~linkedin-profile-scraper:${profile.originalQuery.query}`,
+          profile
+        );
+
+        const member = memberMap[profile.originalQuery.query];
+        const educations = educationMap[member.id];
+        const experiences = experienceMap[member.id];
+
+        let educationCreates = 0;
+        let educationUpdates = 0;
+        let experienceCreates = 0;
+        let experienceUpdates = 0;
+
+        await db.transaction().execute(async (trx) => {
+          await Promise.all([
+            checkMember({ member, profile, trx }),
+
+            ...profile.education.map(async (educationFromLinkedIn) => {
+              const result = await checkEducation({
+                educationFromLinkedIn,
+                educations,
+                memberId: member.id,
+                trx,
+              });
+
+              if (result) {
+                if ('numUpdatedRows' in result) {
+                  educationUpdates += 1;
+                } else {
+                  educationCreates += 1;
+                }
+              }
+            }),
+
+            ...profile.experience.map(async (experienceFromLinkedIn) => {
+              const result = await checkWorkExperience({
+                experienceFromLinkedIn,
+                experiences,
+                memberId: member.id,
+                trx,
+              });
+
+              if (result) {
+                if ('numUpdatedRows' in result) {
+                  experienceUpdates += 1;
+                } else {
+                  experienceCreates += 1;
+                }
+              }
+            }),
+          ]);
+        });
+
+        track({
+          event: 'LinkedIn Synced',
+          properties: {
+            '# of Education Creates': educationCreates,
+            '# of Education Updates': educationUpdates,
+            '# of Work Experience Creates': experienceCreates,
+            '# of Work Experience Updates': experienceUpdates,
+          },
+          user: member.id,
+        });
+      })
+    );
   }
 }
 
-async function getEducationHistoryForDifferential(
-  trx: Transaction<DB>,
-  memberId: string
-) {
+type Member = Awaited<ReturnType<typeof getAllMembers>>[number];
+
+async function getAllMembers(trx: Transaction<DB>, memberIds?: string[]) {
+  return trx
+    .selectFrom('students')
+    .leftJoin('workExperiences', 'workExperiences.studentId', 'students.id')
+    .select([
+      'students.currentLocation',
+      'students.headline',
+      'students.id',
+      'students.linkedInUrl',
+      'students.profilePicture',
+      ({ fn }) => fn.count('workExperiences.id').as('workExperienceCount'),
+    ])
+    .$if(!!memberIds?.length, (qb) => {
+      return qb.where('students.id', 'in', memberIds!);
+    })
+    .$if(!memberIds?.length, (qb) => {
+      return qb
+        .where('students.linkedInUrl', 'is not', null)
+        .where('students.linkedinSyncedAt', 'is', null);
+    })
+    .groupBy('students.id')
+    .orderBy('workExperienceCount', 'asc')
+    .orderBy('students.linkedinSyncedAt', sql`asc nulls first`)
+    .orderBy('acceptedAt', 'asc')
+    .execute();
+}
+
+type Education = Awaited<ReturnType<typeof getAllEducations>>[number];
+
+async function getAllEducations(trx: Transaction<DB>, memberIds?: string[]) {
   return trx
     .selectFrom('educations')
     .leftJoin('schools', 'schools.id', 'educations.schoolId')
@@ -239,725 +261,699 @@ async function getEducationHistoryForDifferential(
       'educations.degreeType',
       'educations.major',
       'educations.otherMajor',
+      'educations.studentId',
+      'schools.linkedinId',
       ({ ref }) => {
-        return sql<string>`to_char(${ref('endDate')}, 'YYYY-MM')`.as('endDate');
+        return sql<string>`to_char(${ref('endDate')}, 'YYYY-MM-DD')`.as(
+          'endDate'
+        );
       },
       ({ ref }) => {
-        return sql<string>`to_char(${ref('startDate')}, 'YYYY-MM')`.as(
+        return sql<string>`to_char(${ref('startDate')}, 'YYYY-MM-DD')`.as(
           'startDate'
         );
       },
       ({ fn }) => {
         return fn
           .coalesce('schools.name', 'educations.otherSchool')
+          .$castTo<string>()
           .as('school');
       },
     ])
-    .where('studentId', '=', memberId)
+    .$if(!!memberIds?.length, (qb) => {
+      return qb.where('educations.studentId', 'in', memberIds!);
+    })
     .where('educations.deletedAt', 'is', null)
     .orderBy('educations.startDate', 'desc')
     .orderBy('educations.endDate', 'desc')
     .execute();
 }
 
-async function getMemberForDifferential(
-  trx: Transaction<DB>,
-  memberId: string
-) {
-  return trx
-    .selectFrom('students')
-    .select(['currentLocation', 'linkedInUrl', 'slackId'])
-    .where('id', '=', memberId)
-    .executeTakeFirst();
-}
+type Experience = Awaited<ReturnType<typeof getAllExperiences>>[number];
 
-async function getWorkHistoryForDifferential(
-  trx: Transaction<DB>,
-  memberId: string
-) {
+async function getAllExperiences(trx: Transaction<DB>, memberIds?: string[]) {
   return trx
     .selectFrom('workExperiences')
     .leftJoin('companies', 'companies.id', 'workExperiences.companyId')
     .select([
+      'companies.linkedinId',
+      'workExperiences.description',
       'workExperiences.employmentType',
       'workExperiences.id',
       'workExperiences.locationCity',
       'workExperiences.locationState',
       'workExperiences.locationType',
+      'workExperiences.studentId',
       'workExperiences.title',
       ({ ref }) => {
-        return sql<string>`to_char(${ref('endDate')}, 'YYYY-MM')`.as('endDate');
+        return sql<string>`to_char(${ref('endDate')}, 'MM')`.as('endMonth');
       },
       ({ ref }) => {
-        return sql<string>`to_char(${ref('startDate')}, 'YYYY-MM')`.as(
+        return sql<string>`to_char(${ref('endDate')}, 'YYYY')`.as('endYear');
+      },
+      ({ ref }) => {
+        return sql<string>`to_char(${ref('endDate')}, 'YYYY-MM-DD')`.as(
+          'endDate'
+        );
+      },
+      ({ ref }) => {
+        return sql<string>`to_char(${ref('startDate')}, 'MM')`.as('startMonth');
+      },
+      ({ ref }) => {
+        return sql<string>`to_char(${ref('startDate')}, 'YYYY')`.as(
+          'startYear'
+        );
+      },
+      ({ ref }) => {
+        return sql<string>`to_char(${ref('startDate')}, 'YYYY-MM-DD')`.as(
           'startDate'
         );
       },
       ({ fn }) => {
         return fn
           .coalesce('companies.name', 'workExperiences.companyName')
+          .$castTo<string>()
           .as('company');
       },
     ])
-    .where('studentId', '=', memberId)
+    .$if(!!memberIds?.length, (qb) => {
+      return qb.where('workExperiences.studentId', 'in', memberIds!);
+    })
     .where('workExperiences.deletedAt', 'is', null)
     .orderBy('workExperiences.endDate', 'desc')
     .orderBy('workExperiences.startDate', 'desc')
     .execute();
 }
 
-const ApifyProfileData = z.object({
-  basic_info: z.object({
-    location: z.object({ full: z.string() }),
-  }),
-  experience: z.unknown().array(),
-  education: z.unknown().array(),
-});
-
-type ApifyProfileData = z.infer<typeof ApifyProfileData>;
-
-/**
- * Gets the LinkedIn profile data for a given LinkedIn profile URL. This
- * function composes other functions to get the LinkedIn profile data:
- *
- * 1. Starts a LinkedIn Profile scraper run in Apify and gets the dataset ID.
- * 2. Uses the dataset ID to get the actual LinkedIn data.
- * 3. Cleans the LinkedIn data with AI.
- * 4. Caches the LinkedIn data for 30 days.
- *
- * Note that the LinkedIn data is cached for 30 days. This is because the
- * LinkedIn Profile scraper run in Apify is rate limited costs roughly $0.05
- * per run and LinkedIn data isn't updated that often.
- *
- * @param url - LinkedIn profile URL to scrape.
- * @returns Promise resolving to the LinkedIn profile data.
- */
-async function getLinkedInProfile(
-  url: string
-): Promise<LinkedInProfile | null> {
-  return withCache(
-    `linkedin:v2:${url}`,
-    60 * 60 * 24 * 30,
-    async function fn() {
-      try {
-        const dataset = await runActor({
-          actorId: 'apimaestro~linkedin-profile-detail',
-          body: { username: url },
-          schema: z.array(ApifyProfileData),
-        });
-
-        const profile = await transformProfileData(dataset[0]);
-
-        return profile;
-      } catch (e) {
-        // There's a lot that can go wrong with the LinkedIn profile scraper, for
-        // example the user's profile is private or they inputted a bad URL. We
-        // don't necessarily want to fail the entire sync process for that, so
-        // we'll just exit gracefully.
-        return null;
-      }
-    }
-  );
-}
-
-// "Clean LinkedIn Profile"
-
-const TRANSFORM_LINKEDIN_PROFILE_PROMPT = dedent`
-  You are given raw LinkedIn profile data. Your task is to extract and transform
-  it into a JSON object that strictly adheres to the schema defined below. Your
-  output **must be valid JSON**, match the schema **exactly**, and include
-  **no extra text**—only the JSON.
-
-  ---
-
-  Schema:
-
-  const LinkedInEducation = z.object({
-    degreeType: z.enum([
-      'associate',
-      'bachelors',
-      'certificate',
-      'doctoral',
-      'masters',
-      'professional',
-    ]),
-    endDate: z.string().nullable(),
-    isSchoolOnLinkedIn: z.boolean(),
-    major: z.enum([
-      'artificial_intelligence',
-      'computer_science',
-      'data_science',
-      'electrical_or_computer_engineering',
-      'information_science',
-      'other',
-    ]),
-    otherMajor: z.string().nullable(),
-    school: z.string(),
-    startDate: z.string().nullable(),
-  });
-
-  const LinkedInExperience = z.object({
-    company: z.string(),
-    endDate: z.string().nullable(),
-    employmentType: z.enum([
-      'apprenticeship',
-      'contract',
-      'freelance',
-      'full_time',
-      'internship',
-      'part_time',
-    ]),
-    isCompanyOnLinkedIn: z.boolean(),
-    location: z.string().nullable(),
-    locationType: z.enum(['hybrid', 'in_person', 'remote']),
-    startDate: z.string(),
-    title: z.string(),
-  });
-
-  const OutputSchema = z.object({
-    educations: z.array(LinkedInEducation),
-    experiences: z.array(LinkedInExperience),
-    location: z.string().nullable(),
-  });
-
-  ---
-
-  Instructions:
-
-  1. Output must be a single valid JSON object matching "OutputSchema".
-  2. Do not include high school or lower education. Only include college-level
-     and above.
-  3. Exclude non-professional and service work (e.g., server, cashier). Include
-     relevant professional experience (e.g., founder, software engineer).
-  4. If "location" is not null and "locationType" is null, set "locationType" to
-     "in_person". If both "location" and "locationType" are null, default
-     "locationType" to "remote".
-  5. Normalize all dates as follows:
-    - Format: "YYYY-MM-DD"
-    - If only year and month → use "YYYY-MM-01"
-    - If only year:
-      - Education start → "YYYY-08-01"
-      - Education end → "YYYY-05-01"
-      - Experience start → "YYYY-01-01"
-      - Experience end → "YYYY-12-01"
-  6. Normalize metro-area LinkedIn locations to city/state as follows:
-    - "San Francisco Bay Area" → "San Francisco, CA"
-    - "New York City Metropolitan Area" → "New York, NY"
-    - "Greater Los Angeles Area" → "Los Angeles, CA"
-    - "Greater Seattle Area" → "Seattle, WA"
-    - "Houston, Texas Area" → "Houston, TX"
-    - "Greater Boston Area" → "Boston, MA"
-    - "Chicago Metropolitan Area" → "Chicago, IL"
-  7. For education majors not listed in the enum, set "major" to "other" and
-     populate "otherMajor" with the raw major.
-
-  ---
-
-  Example:
-
-  Input:
-  """
-  Education:
-  - Degree: Bachelors
-    Major: Computer Science
-    School: MIT
-    Start: 2015
-    End: 2019
-
-  Experience:
-  - Title: Software Engineer
-    Company: Google
-    Start: 2020-03
-    End: Present
-    Employment Type: Full-time
-    Location: San Francisco Bay Area
-    Location Type: Hybrid
-
-  Location: Greater Boston Area
-  """
-
-  Output:
-  {
-    "educations": [
-      {
-        "degreeType": "bachelors",
-        "endDate": "2019-05-01",
-        "isSchoolOnLinkedIn": true,
-        "major": "computer_science",
-        "otherMajor": null,
-        "school": "MIT",
-        "startDate": "2015-08-01"
-      }
-    ],
-    "experiences": [
-      {
-        "company": "Google",
-        "endDate": null,
-        "employmentType": "full_time",
-        "isCompanyOnLinkedIn": true,
-        "location": "San Francisco, CA",
-        "locationType": "hybrid",
-        "startDate": "2020-03-01",
-        "title": "Software Engineer"
-      }
-    ],
-    "location": "Boston, MA"
-  }
-`;
-
-/**
- * Transforms the LinkedIn profile data from Apify into a specific JSON
- * structure. We use AI to follow the schema instructions exactly. We
- * validate the output with Zod.
- *
- * @param profile - LinkedIn profile data from Apify.
- * @returns Promise resolving to the cleaned LinkedIn profile data.
- */
-async function transformProfileData(
-  profile: ApifyProfileData
-): Promise<LinkedInProfile> {
-  profile.experience.forEach((experience: any) => {
-    experience.is_company_on_linkedin =
-      !!experience.company_id &&
-      !!experience.company_linkedin_url &&
-      !!experience.company_logo_url;
-  });
-
-  profile.education.forEach((education: any) => {
-    education.is_school_on_linkedin =
-      !!education.school_id &&
-      !!education.school_linkedin_url &&
-      !!education.school_logo_url;
-  });
-
-  const completionResult = await getChatCompletion({
-    model: 'claude-sonnet-4-20250514',
-    maxTokens: 1000,
-    messages: [
-      {
-        role: 'user',
-        content: dedent`
-          <linkedin_profile>
-            ${JSON.stringify(profile, null, 2)}
-          </linkedin_profile>
-        `,
-      },
-    ],
-    system: [
-      {
-        type: 'text',
-        text: 'You are a data transformation assistant. Output only valid JSON matching the specified schema exactly.',
-      },
-      { type: 'text', text: TRANSFORM_LINKEDIN_PROFILE_PROMPT, cache: true },
-    ],
-    temperature: 0,
-  });
-
-  if (!completionResult.ok) {
-    throw new ColorStackError()
-      .withMessage('Failed to format LinkedIn profile data with AI.')
-      .withContext({ response: completionResult.error })
-      .report();
-  }
-
-  let json: any;
-
-  try {
-    json = JSON.parse(completionResult.data);
-  } catch (error) {
-    throw new ColorStackError()
-      .withMessage('Failed to parse AI response into JSON.')
-      .withContext({ error, response: completionResult.data })
-      .report();
-  }
-
-  const profileResult = LinkedInProfile.safeParse(json);
-
-  if (!profileResult.success) {
-    throw new ColorStackError()
-      .withMessage('Failed to parse LinkedIn Profile from AI.')
-      .withContext({ error: profileResult.error, response: json })
-      .report();
-  }
-
-  return profileResult.data;
-}
-
-const LINKEDIN_DIFFERENTIAL_ROLE = dedent`
-  You are an expert at comparing JSON records. You understand the nuances of the
-  data and can determine when changes are needed.
-`;
-
-const LINKEDIN_DIFFERENTIAL_PROMPT = dedent`
-  <context>
-    Users can manually enter their work history, education, and location, but
-    this data often already exists on LinkedIn. This tool syncs LinkedIn profiles
-    to in-app database profiles.
-  </context>
-
-  <goal>
-    Generate a list of changes to update the user's database profile based on
-    their LinkedIn profile.
-
-    🔁 One-way sync: **LinkedIn → Database**
-    - Add records present in LinkedIn but missing in the database.
-    - Update records if LinkedIn has newer or different data.
-
-    ⚠️ Do **not** generate changes for database records missing from LinkedIn.
-    Only LinkedIn data should drive updates.
-  </goal>
-
-  <types>
-    type EducationChange = {
-      type: 'education';
-      reason: string;
-      data: {
-        degreeType: 'associate' | 'bachelors' | 'certificate' | 'doctoral' |
-          'masters' | 'professional';
-        endDate: string | null;
-        isSchoolOnLinkedIn: boolean;
-        major: 'artificial_intelligence' | 'computer_science' | 'data_science' |
-          'electrical_or_computer_engineering' | 'information_science' | 'other';
-        otherMajor: string | null;
-        school: string;
-        startDate: string | null;
-        id?: string;
-      };
-    }
-
-    type ExperienceChange = {
-      type: 'experience';
-      reason: string;
-      data: {
-        company: string;
-        endDate: string | null;
-        employmentType: 'apprenticeship' | 'contract' | 'freelance' |
-          'full_time' | 'internship' | 'part_time';
-        isCompanyOnLinkedIn: boolean;
-        location: string | null;
-        locationType: 'hybrid' | 'in_person' | 'remote';
-        startDate: string | null;
-        title: string;
-        id?: string;
-      };
-    }
-
-    type LocationChange = {
-      type: 'location';
-      reason: string;
-      data: {
-        location: string;
-      };
-    }
-
-    type Change = EducationChange | ExperienceChange | LocationChange;
-
-    Change[] // Output format
-  </types>
-
-  <rules>
-    - Sync is LinkedIn → Database only.
-    - Ignore database-only data.
-    - Each change must include a clear, concise "reason".
-    - Do not generate changes based on the "isCompanyOnLinkedIn" and
-      "isSchoolOnLinkedIn" fields. They are mainly there for metadata purposes.
-
-    **Date normalization:**
-    - Format: "YYYY-MM-DD"
-    - If only year/month: assume first of month → "2025-01" → "2025-01-01"
-    - If only year:
-      - Education start: "YYYY-08-01"
-      - Education end: "YYYY-05-01"
-      - Experience start: "YYYY-01-01"
-      - Experience end: "YYYY-12-01"
-    - If LinkedIn is less precise (e.g., "2020") than the database (e.g.,
-      "2020-05-01"), do **not** count as a change.
-
-    **Location normalization:**
-    - Map LinkedIn metro areas to city/state:
-      - e.g., "San Francisco Bay Area" → "San Francisco, CA"
-    - Normalize both LinkedIn and database locations to "City, State" before
-      comparing.
-    - Treat "City, State" and "City, State, Country" as equivalent locations.
-      - Example: "San Diego, CA" == "San Diego, CA, USA"
-    - If normalized locations match, do **not** generate a change.
-
-    **Output:**
-    - Return a **valid JSON array** of "Change" objects.
-    - If no changes: return [] (no quotes or comments).
-    - Do **not** include:
-      - Markdown formatting (like \`\`\`json)
-      - Comments or explanations
-      - Text before or after the array
-  </rules>
-`;
-
-type ProfileData = {
-  educations: Array<object>;
-  location: string | null;
-  workExperiences: Array<object>;
+type CheckMemberInput = {
+  member: Member;
+  profile: LinkedInProfile;
+  trx: Transaction<DB>;
 };
 
-/**
- * Gets the differential between the LinkedIn profile and the database profile.
- * The result of this function is a list of changes that when executed, will
- * synchronize the database with the LinkedIn profile data.
- *
- * This function relies on AI to evaluate the differences between what's
- * currently in our database on what's on LinkedIn. If there are no differences,
- * then this function SHOULD return an empty array.
- *
- * @param memberId - ID of the member to get the differential for.
- * @returns Promise resolving to the list of changes to synchronize the database
- * with the LinkedIn profile data.
- */
-export async function getProfileDifferential(
-  currentProfile: ProfileData,
-  linkedInProfile: ProfileData
-): Promise<Change[]> {
-  const content = dedent`
-    Here is the data from the database and the LinkedIn profile:
+async function checkMember({ member, profile, trx }: CheckMemberInput) {
+  const updatedLocation = await run(async () => {
+    const locationFromLinkedIn = profile.element.location.parsed.text;
 
-    <database_profile>
-      ${JSON.stringify(currentProfile, null, 2)}
-    </database_profile>
-
-    <linkedin_profile>
-      ${JSON.stringify(linkedInProfile, null, 2)}
-    </linkedin_profile>
-  `;
-
-  const completionResult = await getChatCompletion({
-    model: 'claude-sonnet-4-20250514',
-    maxTokens: 1000,
-    messages: [{ role: 'user', content }],
-    system: [
-      { type: 'text', text: LINKEDIN_DIFFERENTIAL_ROLE, cache: true },
-      { type: 'text', text: LINKEDIN_DIFFERENTIAL_PROMPT, cache: true },
-    ],
-    temperature: 0,
+    if (
+      locationFromLinkedIn &&
+      locationFromLinkedIn !== member.currentLocation
+    ) {
+      return getMostRelevantLocation(locationFromLinkedIn);
+    }
   });
-
-  if (!completionResult.ok) {
-    throw new ColorStackError()
-      .withMessage('Failed to generate LinkedIn sync changes.')
-      .withContext({ response: completionResult.error })
-      .report();
-  }
-
-  let json: any;
-
-  try {
-    json = JSON.parse(completionResult.data);
-  } catch (error) {
-    throw new ColorStackError()
-      .withMessage('Failed to parse LinkedIn sync changes from AI.')
-      .withContext({ error, response: completionResult.data })
-      .report();
-  }
-
-  const changesResult = Change.array().safeParse(json);
-
-  if (!changesResult.success) {
-    throw new ColorStackError()
-      .withMessage('Failed to parse LinkedIn sync changes from AI.')
-      .withContext({ error: changesResult.error, response: json })
-      .report();
-  }
-
-  return changesResult.data;
-}
-
-/**
- * Upserts an education record from LinkedIn profile data. If the education
- * already exists, it will be updated. If it does not exist, it will be created.
- *
- * @param trx - Database transaction to use for the query.
- * @param memberId - ID of the member to upsert the education for.
- * @param data - LinkedIn education data to upsert.
- */
-async function upsertEducationFromLinkedIn(
-  trx: Transaction<DB>,
-  memberId: string,
-  data: Extract<Change, { type: 'education' }>['data']
-) {
-  const educationId = data.id || id();
-
-  const school = data.isSchoolOnLinkedIn
-    ? await getMostRelevantSchool(trx, data.school)
-    : null;
-
-  return trx
-    .insertInto('educations')
-    .values({
-      ...(school
-        ? { otherSchool: null, schoolId: school.id }
-        : { otherSchool: data.school, schoolId: null }),
-
-      degreeType: data.degreeType,
-      endDate: data.endDate,
-      id: educationId,
-      major: data.major,
-      otherMajor: data.otherMajor,
-      startDate: data.startDate,
-      studentId: memberId,
-    })
-    .onConflict((oc) => {
-      return oc.column('id').doUpdateSet((eb) => {
-        return {
-          degreeType: eb.ref('excluded.degreeType'),
-          endDate: eb.ref('excluded.endDate'),
-          major: eb.ref('excluded.major'),
-          otherMajor: eb.ref('excluded.otherMajor'),
-          otherSchool: eb.ref('excluded.otherSchool'),
-          startDate: eb.ref('excluded.startDate'),
-          schoolId: eb.ref('excluded.schoolId'),
-          updatedAt: new Date(),
-        };
-      });
-    })
-    .execute();
-}
-
-/**
- * Finds the most relevant school in the database matching a school name.
- * Uses fuzzy text matching via PostgreSQL similarity functions to find
- * closest matches.
- *
- * This is useful because LinkedIn will just have a raw school name for each
- * education, and we need to attempt to match it to our database of schools.
- *
- * @param trx - Database transaction to use for the query.
- * @param schoolName - Name of school to search for.
- * @returns Promise resolving to the ID of the most relevant matching school,
- * if found.
- */
-async function getMostRelevantSchool(trx: Transaction<DB>, schoolName: string) {
-  const similarity = sql`similarity(name, ${schoolName})`;
-  const wordSimilarity = sql`word_similarity(name, ${schoolName})`;
-
-  return trx
-    .selectFrom('schools')
-    .select('id')
-    .where((eb) => {
-      return eb.or([
-        eb('name', 'ilike', `%${schoolName}%`),
-        eb(similarity, '>', 0.5),
-        eb(wordSimilarity, '>', 0.5),
-      ]);
-    })
-    .orderBy(similarity, 'desc')
-    .orderBy(wordSimilarity, 'desc')
-    .limit(1)
-    .executeTakeFirst();
-}
-
-/**
- * Upserts a work experience record from LinkedIn profile data. If the work
- * experience already exists, it will be updated. If it does not exist, it will
- * be created.
- *
- * @param trx - Database transaction to use for the query.
- * @param memberId - ID of the member to upsert the work experience for.
- * @param data - LinkedIn work experience data to upsert.
- */
-async function upsertExperienceFromLinkedIn(
-  trx: Transaction<DB>,
-  memberId: string,
-  data: Extract<Change, { type: 'experience' }>['data']
-) {
-  const workExperienceId = data.id || id();
-
-  // `data.company` is the raw company name from LinkedIn. We need to find the
-  // most relevant company in our database or using the Crunchbase API.
-  const companyId = data.isCompanyOnLinkedIn
-    ? await saveCompanyIfNecessary(trx, data.company)
-    : null;
-
-  const location = await getMostRelevantLocation(data.location);
-
-  let locationType: LocationType;
-
-  if (data.locationType) {
-    locationType = data.locationType;
-  } else if (location) {
-    // If there is a location found, we'll assume it's in-person.
-    locationType = LocationType.IN_PERSON;
-  } else {
-    // If there is no location found, we'll assume it's remote.
-    locationType = LocationType.REMOTE;
-  }
-
-  return trx
-    .insertInto('workExperiences')
-    .values({
-      ...(companyId
-        ? { companyId, companyName: null }
-        : { companyId: null, companyName: data.company }),
-
-      ...(location
-        ? { locationCity: location.city, locationState: location.state }
-        : { locationCity: null, locationState: null }),
-
-      employmentType: data.employmentType,
-      endDate: data.endDate,
-      id: workExperienceId,
-      locationType,
-      startDate: data.startDate,
-      studentId: memberId,
-      title: data.title,
-    })
-    .onConflict((oc) => {
-      return oc.column('id').doUpdateSet(({ ref }) => {
-        return {
-          companyId: ref('excluded.companyId'),
-          companyName: ref('excluded.companyName'),
-          employmentType: ref('excluded.employmentType'),
-          endDate: ref('excluded.endDate'),
-          locationCity: ref('excluded.locationCity'),
-          locationState: ref('excluded.locationState'),
-          locationType: ref('excluded.locationType'),
-          startDate: ref('excluded.startDate'),
-          studentId: ref('excluded.studentId'),
-          title: ref('excluded.title'),
-          updatedAt: new Date(),
-        };
-      });
-    })
-    .execute();
-}
-
-/**
- * Updates a member's current location from LinkedIn profile data. We'll
- * default to their location on LinkedIn. If we can't find a match using the
- * Google Places API, we won't update anything.
- *
- * @param trx - Database transaction to use for the query.
- * @param memberId - ID of the member to update the location for.
- * @param data - LinkedIn location data to update.
- * @returns Promise resolving to the result of the update.
- */
-async function updateLocationFromLinkedIn(
-  trx: Transaction<DB>,
-  memberId: string,
-  data: Extract<Change, { type: 'location' }>['data']
-) {
-  const location = await getMostRelevantLocation(data.location);
-
-  if (!location) {
-    return null;
-  }
 
   return trx
     .updateTable('students')
     .set({
-      currentLocation: location.formattedAddress,
-      currentLocationCoordinates: point({
-        x: location.longitude,
-        y: location.latitude,
+      ...(!member.headline && {
+        headline: profile.element.headline,
+      }),
+      ...(!member.profilePicture && {
+        profilePicture: profile.element.photo,
+      }),
+      ...(updatedLocation && {
+        currentLocation: updatedLocation.formattedAddress,
+        currentLocationCoordinates: point({
+          x: updatedLocation.longitude,
+          y: updatedLocation.latitude,
+        }),
+      }),
+      linkedinSyncedAt: new Date(),
+      updatedAt: new Date(),
+    })
+    .where('id', '=', member.id)
+    .executeTakeFirst();
+}
+
+const MONTH_MAP: Record<string, string> = {
+  Jan: '01',
+  Feb: '02',
+  Mar: '03',
+  Apr: '04',
+  May: '05',
+  Jun: '06',
+  Jul: '07',
+  Aug: '08',
+  Sep: '09',
+  Oct: '10',
+  Nov: '11',
+  Dec: '12',
+};
+
+type CheckEducationInput = {
+  educationFromLinkedIn: LinkedInEducation;
+  educations: Education[];
+  memberId: string;
+  trx: Transaction<DB>;
+};
+
+async function checkEducation({
+  educationFromLinkedIn,
+  educations,
+  memberId,
+  trx,
+}: CheckEducationInput) {
+  // If there is no `degree` field, it likely means that the education was
+  // pre-college. If there is no `fieldOfStudy` field, it likely means that
+  // the education was part of a general education program. We also will require
+  // that the date fields are present.
+  if (
+    !educationFromLinkedIn.degree ||
+    !educationFromLinkedIn.fieldOfStudy ||
+    !educationFromLinkedIn.endDate ||
+    !educationFromLinkedIn.startDate
+  ) {
+    return;
+  }
+
+  const degreeType = getDegreeType(educationFromLinkedIn.degree);
+
+  // If we don't have a matching degree type, we'll skip.
+  if (!degreeType) {
+    return;
+  }
+
+  const url = new URL(educationFromLinkedIn.schoolLinkedinUrl);
+
+  // Example: https://www.linkedin.com/company/123 -> 123
+  // Example: https://www.linkedin.com/school/123 -> 123
+  const linkedinId = url.pathname.split('/').filter(Boolean)[1];
+
+  // If there is no ID, it likely means that the school is not accredited so
+  // we will skip syncing it.
+  if (!linkedinId) {
+    return;
+  }
+
+  const existingEducation = educations.find((education) => {
+    if (education.degreeType !== degreeType) {
+      return false;
+    }
+
+    // Usually we would just check for the exact ID match, but sometimes a
+    // member on LinkedIn will link to a different account (ie: Cornell
+    // University vs. Cornell University Engineering). So we'll check for
+    // a close match from the school name as well.
+    return (
+      education.linkedinId === linkedinId ||
+      education.school.includes(educationFromLinkedIn.schoolName) ||
+      educationFromLinkedIn.schoolName.includes(education.school)
+    );
+  });
+
+  const startDate = run(() => {
+    const startMonth = educationFromLinkedIn.startDate!.month;
+    const startYear = educationFromLinkedIn.startDate!.year;
+
+    if (startYear && startMonth) {
+      // The `startMonth` is formatted as an abbreviation of the month, so
+      // we need to convert it to a 2-digit number (ie: `Jan` -> `01`).
+      return `${startYear}-${MONTH_MAP[startMonth]}-01`;
+    } else if (startYear) {
+      // If there is no `startMonth`, we'll default to August since that's
+      // when most schools start their academic year.
+      return `${startYear}-08-01`;
+    }
+
+    return undefined;
+  });
+
+  const endDate = run(() => {
+    const endMonth = educationFromLinkedIn.endDate!.month;
+    const endYear = educationFromLinkedIn.endDate!.year;
+
+    if (endMonth && endYear) {
+      // The `endMonth` is formatted as an abbreviation of the month, so
+      // we need to convert it to a 2-digit number (ie: `Jan` -> `01`).
+      return `${endYear}-${MONTH_MAP[endMonth]}-01`;
+    } else if (endYear) {
+      // If there is no `endMonth`, we'll default to May since that's
+      // when most schools end their academic year.
+      return `${endYear}-05-01`;
+    }
+
+    return undefined;
+  });
+
+  if (existingEducation) {
+    const set: Updateable<DB['educations']> = {
+      updatedAt: new Date(),
+    };
+
+    if (existingEducation.linkedinId !== linkedinId) {
+      const schoolId = await saveSchoolIfNecessary(trx, linkedinId);
+
+      if (schoolId) {
+        set.schoolId = schoolId;
+        set.otherSchool = null;
+      } else {
+        set.otherSchool = educationFromLinkedIn.schoolName;
+        set.schoolId = null;
+      }
+    }
+
+    if (existingEducation.endDate !== endDate) {
+      set.endDate = endDate;
+    }
+
+    if (existingEducation.startDate !== startDate) {
+      set.startDate = startDate;
+    }
+
+    if (!Object.keys(set).length) {
+      return;
+    }
+
+    return trx
+      .updateTable('educations')
+      .set(set)
+      .where('id', '=', existingEducation.id)
+      .executeTakeFirst();
+  }
+
+  const schoolId = await saveSchoolIfNecessary(trx, linkedinId);
+
+  const major = getFieldOfStudy(educationFromLinkedIn.fieldOfStudy);
+
+  return trx
+    .insertInto('educations')
+    .values({
+      createdAt: new Date(),
+      degreeType,
+      endDate,
+      id: id(),
+      startDate,
+      studentId: memberId,
+      updatedAt: new Date(),
+
+      ...(schoolId
+        ? { schoolId }
+        : { otherSchool: educationFromLinkedIn.schoolName }),
+
+      ...(major
+        ? { major }
+        : { major: 'other', otherMajor: educationFromLinkedIn.fieldOfStudy }),
+    })
+    .executeTakeFirst();
+}
+
+/**
+ * Converts a LinkedIn degree type to an Oyster degree type. This is a best
+ * effort conversion and may not be 100% accurate.
+ *
+ * @param degree - LinkedIn degree type.
+ * @returns Oyster degree type or `certificate` if the degree is not recognized.
+ */
+function getDegreeType(degree: string) {
+  const value = degree.toLowerCase();
+
+  if (value.includes('bachelor')) {
+    return 'bachelors';
+  }
+
+  if (value.includes('master')) {
+    return 'masters';
+  }
+
+  if (value.includes('doctor')) {
+    return 'doctoral';
+  }
+
+  if (value.includes('associate')) {
+    return 'associate';
+  }
+
+  if (value.includes('certificate')) {
+    return 'certificate';
+  }
+
+  return null;
+}
+
+/**
+ * Converts a LinkedIn field of study to an Oyster major. This is a best
+ * effort conversion and may not be 100% accurate.
+ *
+ * @param fieldOfStudy - LinkedIn field of study.
+ * @returns Oyster major or `null` if the field of study is not recognized.
+ */
+function getFieldOfStudy(fieldOfStudy: string): Major | null {
+  if (fieldOfStudy.includes('Computer Science')) {
+    return 'computer_science';
+  }
+
+  if (
+    fieldOfStudy.includes('Computer Engineering') ||
+    fieldOfStudy.includes('Electrical Engineering')
+  ) {
+    return 'electrical_or_computer_engineering';
+  }
+
+  if (
+    fieldOfStudy.includes('Data Science') ||
+    fieldOfStudy.includes('Data Analytics')
+  ) {
+    return 'data_science';
+  }
+
+  if (fieldOfStudy.includes('Information Science')) {
+    return 'information_science';
+  }
+
+  if (fieldOfStudy.includes('Software Engineering')) {
+    return 'software_engineering';
+  }
+
+  if (fieldOfStudy.includes('Cybersecurity')) {
+    return 'cybersecurity';
+  }
+
+  return null;
+}
+
+// CHECKPOINT
+
+type CheckWorkExperienceInput = {
+  experienceFromLinkedIn: LinkedInExperience;
+  experiences: Experience[];
+  memberId: string;
+  trx: Transaction<DB>;
+};
+
+async function checkWorkExperience({
+  experienceFromLinkedIn,
+  experiences,
+  memberId,
+  trx,
+}: CheckWorkExperienceInput) {
+  // We're going to ignore any work experiences that don't have a company ID
+  // or start date for now...this errs on the side of caution.
+  if (
+    !experienceFromLinkedIn.companyId ||
+    !experienceFromLinkedIn.startDate ||
+    !experienceFromLinkedIn.startDate.year
+  ) {
+    return;
+  }
+
+  const existingExperience = experiences.find((experience) => {
+    return doesExperienceMatch(experienceFromLinkedIn, experience);
+  });
+
+  if (existingExperience) {
+    return updateWorkExperience({
+      existingExperience,
+      experienceFromLinkedIn,
+      trx,
+    });
+  }
+
+  return createWorkExperience({
+    experienceFromLinkedIn,
+    memberId,
+    trx,
+  });
+}
+
+function doesExperienceMatch(
+  experienceFromLinkedIn: LinkedInExperience,
+  experience: Experience
+): boolean {
+  // We allow for close matches to happen here simply to reduce the amount of
+  // false positives (would rather update something than create a duplicate).
+  if (
+    experience.linkedinId !== experienceFromLinkedIn.companyId &&
+    !experience.company.includes(experienceFromLinkedIn.companyName) &&
+    !experienceFromLinkedIn.companyName.includes(experience.company)
+  ) {
+    return false;
+  }
+
+  // Now, we'll implement a "confidence" score to determine if the new
+  // experience is the same as the existing experience.
+
+  let score = 0;
+
+  // The position name is the most important factor so if it's an exact match
+  // then we'll give it a score of 2.
+  if (experienceFromLinkedIn.position === experience.title) {
+    score += 2;
+  }
+
+  if (
+    experienceFromLinkedIn.startDate?.year &&
+    experienceFromLinkedIn.startDate?.year === parseInt(experience.startYear)
+  ) {
+    score++;
+  }
+
+  if (
+    experienceFromLinkedIn.startDate?.month &&
+    MONTH_MAP[experienceFromLinkedIn.startDate.month] === experience.startMonth
+  ) {
+    score++;
+  }
+
+  if (
+    experience.endYear &&
+    experienceFromLinkedIn.endDate?.year &&
+    experienceFromLinkedIn.endDate?.year === parseInt(experience.endYear)
+  ) {
+    score++;
+  }
+
+  if (
+    experience.endMonth &&
+    experienceFromLinkedIn.endDate?.month &&
+    MONTH_MAP[experienceFromLinkedIn.endDate.month] === experience.endMonth
+  ) {
+    score++;
+  }
+
+  if (
+    !experience.endYear &&
+    !experience.endMonth &&
+    !experienceFromLinkedIn.endDate
+  ) {
+    score++;
+  }
+
+  return score >= 3;
+}
+
+const EMPLOYMENT_TYPE_MAP: Record<string, string> = {
+  Apprenticeship: EmploymentType.APPRENTICESHIP,
+  Contract: EmploymentType.CONTRACT,
+  Freelance: EmploymentType.FREELANCE,
+  'Full-time': EmploymentType.FULL_TIME,
+  Internship: EmploymentType.INTERNSHIP,
+  'Part-time': EmploymentType.PART_TIME,
+};
+
+const LOCATION_TYPE_MAP: Record<string, string> = {
+  Hybrid: LocationType.HYBRID,
+  'On-site': LocationType.IN_PERSON,
+  Remote: LocationType.REMOTE,
+};
+
+type CreateWorkExperienceInput = {
+  experienceFromLinkedIn: LinkedInExperience;
+  memberId: string;
+  trx: Transaction<DB>;
+};
+
+async function createWorkExperience({
+  experienceFromLinkedIn,
+  memberId,
+  trx,
+}: CreateWorkExperienceInput) {
+  const startDate = run(() => {
+    const startMonth = experienceFromLinkedIn.startDate?.month;
+    const startYear = experienceFromLinkedIn.startDate?.year;
+
+    if (startYear && startMonth) {
+      // The `startMonth` is formatted as an abbreviation of the month, so
+      // we need to convert it to a 2-digit number (ie: `Jan` -> `01`).
+      return `${startYear}-${MONTH_MAP[startMonth]}-01`;
+    }
+
+    return `${startYear}-01-01`;
+  });
+
+  const endDate = run(() => {
+    const endMonth = experienceFromLinkedIn.endDate?.month;
+    const endYear = experienceFromLinkedIn.endDate?.year;
+
+    if (endMonth && endYear) {
+      // The `endMonth` is formatted as an abbreviation of the month, so
+      // we need to convert it to a 2-digit number (ie: `Jan` -> `01`).
+      return `${endYear}-${MONTH_MAP[endMonth]}-01`;
+    } else if (endYear) {
+      // If there is no `endMonth`, we'll default to December since that's
+      // when most companies end their fiscal year.
+      return `${endYear}-12-01`;
+    }
+
+    return undefined;
+  });
+
+  const [companyId, location] = await Promise.all([
+    saveCompanyIfNecessary(trx, experienceFromLinkedIn.companyId),
+    getMostRelevantLocation(experienceFromLinkedIn.location),
+  ]);
+
+  return trx
+    .insertInto('workExperiences')
+    .values({
+      description: experienceFromLinkedIn.description,
+      endDate,
+      id: id(),
+      linkedinSyncedAt: new Date(),
+      startDate,
+      studentId: memberId,
+      title: experienceFromLinkedIn.position,
+
+      ...(companyId
+        ? { companyId }
+        : { companyName: experienceFromLinkedIn.companyName }),
+
+      ...(location && {
+        locationCity: location.city,
+        locationState: location.state,
+      }),
+
+      ...(experienceFromLinkedIn.employmentType && {
+        employmentType:
+          EMPLOYMENT_TYPE_MAP[experienceFromLinkedIn.employmentType],
+      }),
+
+      ...(experienceFromLinkedIn.workplaceType && {
+        locationType: LOCATION_TYPE_MAP[experienceFromLinkedIn.workplaceType],
       }),
     })
-    .where('id', '=', memberId)
+    .execute();
+}
+
+type UpdateWorkExperienceInput = {
+  existingExperience: Experience;
+  experienceFromLinkedIn: LinkedInExperience;
+  trx: Transaction<DB>;
+};
+
+async function updateWorkExperience({
+  existingExperience,
+  experienceFromLinkedIn,
+  trx,
+}: UpdateWorkExperienceInput) {
+  const set: Updateable<DB['workExperiences']> = {
+    linkedinSyncedAt: new Date(),
+    updatedAt: new Date(),
+  };
+
+  if (existingExperience.linkedinId !== experienceFromLinkedIn.companyId) {
+    const companyId = await saveCompanyIfNecessary(
+      trx,
+      experienceFromLinkedIn.companyId
+    );
+
+    if (companyId) {
+      set.companyId = companyId;
+      set.companyName = null;
+    } else {
+      set.companyId = null;
+      set.companyName = experienceFromLinkedIn.companyName;
+    }
+  }
+
+  if (existingExperience.title !== experienceFromLinkedIn.position) {
+    set.title = experienceFromLinkedIn.position;
+  }
+
+  if (existingExperience.description !== experienceFromLinkedIn.description) {
+    set.description = experienceFromLinkedIn.description;
+  }
+
+  const employmentType =
+    EMPLOYMENT_TYPE_MAP[experienceFromLinkedIn.employmentType || ''];
+
+  if (employmentType && existingExperience.employmentType !== employmentType) {
+    set.employmentType = employmentType;
+  }
+
+  if (experienceFromLinkedIn.location) {
+    const location = await getMostRelevantLocation(
+      experienceFromLinkedIn.location
+    );
+
+    if (location) {
+      if (
+        existingExperience.locationCity !== location.city ||
+        existingExperience.locationState !== location.state
+      ) {
+        set.locationCity = location.city;
+        set.locationState = location.state;
+      }
+    }
+  }
+
+  const locationType =
+    LOCATION_TYPE_MAP[experienceFromLinkedIn.workplaceType || ''];
+
+  if (locationType && existingExperience.locationType !== locationType) {
+    set.locationType = locationType;
+  }
+
+  const startDate = run(() => {
+    const startMonth = experienceFromLinkedIn.startDate?.month;
+    const startYear = experienceFromLinkedIn.startDate?.year;
+
+    if (startYear && startMonth) {
+      // The `startMonth` is formatted as an abbreviation of the month, so
+      // we need to convert it to a 2-digit number (ie: `Jan` -> `01`).
+      return `${startYear}-${MONTH_MAP[startMonth]}-01`;
+    } else if (startYear) {
+      // If there is no `startMonth`, we'll default to January since that's
+      // when most companies start their fiscal year.
+      return `${startYear}-01-01`;
+    }
+
+    return undefined;
+  });
+
+  if (startDate && existingExperience.startDate !== startDate) {
+    set.startDate = startDate;
+  }
+
+  const endDate = run(() => {
+    const endMonth = experienceFromLinkedIn.endDate?.month;
+    const endYear = experienceFromLinkedIn.endDate?.year;
+
+    if (endMonth && endYear) {
+      // The `endMonth` is formatted as an abbreviation of the month, so
+      // we need to convert it to a 2-digit number (ie: `Jan` -> `01`).
+      return `${endYear}-${MONTH_MAP[endMonth]}-01`;
+    } else if (endYear) {
+      // If there is no `endMonth`, we'll default to December since that's
+      // when most companies end their fiscal year.
+      return `${endYear}-12-01`;
+    }
+
+    return undefined;
+  });
+
+  if (endDate && existingExperience.endDate !== endDate) {
+    set.endDate = endDate;
+  }
+
+  if (!Object.keys(set).length) {
+    return;
+  }
+
+  return trx
+    .updateTable('workExperiences')
+    .set(set)
+    .where('id', '=', existingExperience.id)
     .execute();
 }
 
@@ -970,18 +966,28 @@ async function updateLocationFromLinkedIn(
  * @param location - Location name to search for.
  * @returns Promise resolving to the most relevant matching location, if found.
  */
-async function getMostRelevantLocation(location: string | null) {
-  if (location) {
-    const cities = await getAutocompletedCities(location);
+async function getMostRelevantLocation(location: string | null | undefined) {
+  if (!location) {
+    return null;
+  }
 
-    if (cities?.length) {
+  return withCache(
+    `${getMostRelevantLocation.name}:${location}`,
+    ONE_WEEK_IN_SECONDS * 4,
+    async function fn() {
+      const cities = await getAutocompletedCities(location);
+
+      if (!cities || !cities.length) {
+        return null;
+      }
+
       const details = await getCityDetails(cities[0].id);
 
       if (details && details.city && details.state) {
         return details;
       }
-    }
-  }
 
-  return null;
+      return null;
+    }
+  );
 }
