+++ packages/core/src/modules/members/members.worker.ts
@@ -14,7 +14,7 @@ import {
   AIRTABLE_MEMBERS_TABLE_ID,
 } from '@/modules/airtable';
 import { sendCompanyReviewNotifications } from '@/modules/employment/use-cases/send-company-review-notifications';
-import { syncLinkedInProfile } from '@/modules/linkedin';
+import { syncLinkedInProfiles } from '@/modules/linkedin';
 import { sendAnniversaryEmail } from '@/modules/members/use-cases/send-anniversary-email';
 import { sendGraduationEmail } from '@/modules/members/use-cases/send-graduation-email';
 import { success } from '@/shared/utils/core';
@@ -59,7 +59,7 @@ export const memberWorker = registerWorker(
         return sendGraduationEmail(data);
       })
       .with({ name: 'student.linkedin.sync' }, ({ data }) => {
-        return syncLinkedInProfile(data.studentId);
+        return syncLinkedInProfiles(data.memberIds);
       })
       .with({ name: 'student.points.recurring' }, ({ data: _ }) => {
         return updatePointTotals();
